+++ portal/views/home.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 # Code for Life
 #
-# Copyright (C) 2019, Ocado Innovation Limited
+# Copyright (C) 2021, Ocado Innovation Limited
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU Affero General Public License as
@@ -40,22 +40,17 @@ from common.helpers.emails import (
     send_email,
     NOTIFICATION_EMAIL,
     add_to_dotmailer,
-    get_dotmailer_user_by_email,
-    send_dotmailer_consent_confirmation_email_to_user,
-    add_consent_record_to_dotmailer_user,
 )
 from common.models import Teacher, Student
 from common.permissions import logged_in_as_student, logged_in_as_teacher
 from django.contrib import messages as messages
 from django.contrib.auth import logout
-from django.http import HttpResponseRedirect, HttpResponse
+from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse, reverse_lazy
 from django.views.decorators.cache import cache_control
-from django.views.decorators.csrf import csrf_exempt
 
 from deploy import captcha
-from portal.forms.dotmailer import NewsletterForm, ConsentForm
 from portal.forms.play import IndependentStudentSignupForm
 from portal.forms.teach import TeacherSignupForm
 from portal.helpers.captcha import remove_captcha_from_forms
@@ -277,53 +272,6 @@ def redirect_teacher_to_correct_page(request, teacher):
             return reverse_lazy("onboarding-classes")
     else:
         return reverse_lazy("onboarding-organisation")
-
-
-@csrf_exempt
-def process_newsletter_form(request):
-    if request.method == "POST":
-        newsletter_form = NewsletterForm(data=request.POST)
-        if newsletter_form.is_valid():
-            user_email = newsletter_form.cleaned_data["email"]
-            add_to_dotmailer("", "", user_email)
-            messages.success(request, "Thank you for signing up! 🎉")
-            return HttpResponseRedirect(reverse_lazy("home"))
-        messages.error(
-            request,
-            "Invalid email address. Please try again.",
-            extra_tags="sub-nav--warning",
-        )
-        return HttpResponseRedirect(reverse_lazy("home"))
-
-    return HttpResponse(status=405)
-
-
-def dotmailer_consent_form(request):
-    if request.method == "POST":
-        consent_form = ConsentForm(data=request.POST)
-
-        if consent_form.is_valid():
-            user_email = consent_form.cleaned_data["email"]
-            user = get_dotmailer_user_by_email(user_email)
-            add_consent_record_to_dotmailer_user(user)
-            send_dotmailer_consent_confirmation_email_to_user(user)
-            return HttpResponseRedirect(reverse_lazy("home"))
-
-        messages.error(
-            request,
-            "Valid email address and consent required. Please try again.",
-            extra_tags="sub-nav--warning",
-        )
-        return HttpResponseRedirect(reverse_lazy("consent_form"))
-
-    else:
-        consent_form = ConsentForm()
-
-    return render(
-        request,
-        "portal/dotmailer_consent_form.html",
-        {"form": consent_form},
-    )
 
 
 @cache_control(private=True)
